# Test optimizations