--- a/tests/test_end_to_end.py
+++ b/tests/test_end_to_end.py
@@ -55,11 +55,31 @@ def test_model_info(client):
     ]
 
 
-def test_http(client):
-    response = client.get("/species/{}".format('ECOLX'))
+def test_species(client):
+    response = client.get("/species/ECOLX")
     assert response.status_code == 200
+    assert len(response.json) > 0
+
+
+def test_simulate_wrong_id(client):
+    response = client.post("/models/wrong_id/simulate", json={'message': {}})
+    assert response.status_code == 404
+
+
+def test_simulate_no_message(client):
+    response = client.post("/models/iJO1366/simulate", json={})
+    assert response.status_code == 400
+
+
+def test_simulate_infeasible(client):
+    response = client.post("/models/iJO1366/simulate", json={'message': MESSAGE_FLUXES_INFEASIBLE})
+    result = response.json
+    assert result['fluxes']['ATPM'] == pytest.approx(100)
+
+
+def test_simulate_modify(client):
     for query, message in {'modify': MESSAGE_MODIFY, 'fluxes': MESSAGE_FLUXES}.items():
-        response = client.post("/models/{}/simulate".format('iJO1366'), json={'message': message})
+        response = client.post("/models/iJO1366/simulate", json={'message': message})
         assert response.status_code == 200
         model = response.json
         assert set(model.keys()) == set(message['to-return'] + ['model-id'])
@@ -75,21 +95,16 @@ def test_http(client):
             assert 'PFK' in {rxn['id'] for rxn in changes['measured']['reactions']}
             assert 'b2297' in {rxn['id'] for rxn in changes['removed']['genes']}
             assert 'BAD_ID' in {rxn['id'] for rxn in changes['measured-missing']['reactions']}
-    response = client.post("/models/{}/simulate".format('iJO1366'), json={'message': MESSAGE_FLUXES_INFEASIBLE})
-    model = response.json
-    assert model['fluxes']['ATPM'] == pytest.approx(100)
-    response = client.post("/models/{}/simulate".format('wrong_id'), json={'message': {}})
-    assert response.status_code == 404
-    response = client.post("/models/{}/simulate".format('iJO1366'), json={})
-    assert response.status_code == 400
-    response_etoh = client.post("/models/{}/simulate".format('iJO1366'),
-                                json={'message': MESSAGE_DIFFERENT_OBJECTIVE})
+
+
+def test_simulate_different_objective(client):
+    response_etoh = client.post("/models/iJO1366/simulate", json={'message': MESSAGE_DIFFERENT_OBJECTIVE})
     assert response_etoh.status_code == 200
-    model = response_etoh.json
-    assert abs(model['fluxes']['EX_etoh_e']) - 20.0 < 0.001
+    result = response_etoh.json
+    assert abs(result['fluxes']['EX_etoh_e']) - 20.0 < 0.001
+
     MESSAGE_DIFFERENT_OBJECTIVE.pop('objective')
-    response_etoh = client.post("/models/{}/simulate".format('iJO1366'),
-                                json={'message': MESSAGE_DIFFERENT_OBJECTIVE})
+    response_etoh = client.post("/models/iJO1366/simulate", json={'message': MESSAGE_DIFFERENT_OBJECTIVE})
     assert response_etoh.status_code == 200
-    model = response_etoh.json
-    assert abs(model['fluxes']['EX_etoh_e']) < 0.001
+    result = response_etoh.json
+    assert abs(result['fluxes']['EX_etoh_e']) < 0.001