Based on `flask` -> [Storage refactor](https://github.com/DD-DeCaF/model/pull/106) -> [Simulations refactor](https://github.com/DD-DeCaF/model/pull/107)

- Provide modifiable fixtures of models used in test cases. Uses the cobrapy context manager for performance.
- API queries to ICE and id-mapper are mocked where integration isn't explicitly tested
- Use the slim `e_coli_core` model where possible, for efficiency
- Add MetanetX universal bigg model instead for loading it through cameo. (Decreases load time from ~20s to ~6s)

In total, these optimizations decrease the test suite runtime from ~554s to ~92s.