@kvikshaug 2018-07-20
This was merged in https://github.com/DD-DeCaF/model/commit/ceedd1870a062e92d96245ff5b460e77d686a3bb, but due to the Github service interruption on July 16, the post_receive hooks hasn't run successfully, so I'm closing the PR.